Use neutral arguments in `Vault.create_hash` signature
Currently, when generating a passphrase from an SSH key signature,
`derivepassphrase.Vault.create_hash` requires the user to know the role
of the `key` and `message` arguments themselves; in particular, it
requires the user to add the vault UUID to the message themselves.  The
fact that the vault UUID is added to the hashing input is an
implementation detail, as should be the fact that one of the arguments
is a key and the other is a salt value to a key-derivation function. Put
these details only into the docstring's description section, not the
argument names; rename them to `phrase` and `service`, respectively.