Fortify the argument parsing and handling in the command-line interface
Currently, we use a one-to-one correspondence between command-line
argument name and function argument, via catch-all keyword arguments.
This is very unwieldy to use with type checking, it exposes us to name
clashes with Python keywords, and it doesn't even save that much typing
effort because the arguments still ought to be documented in the
documentation anyway. Therefore, introduce an explicit target parameter
name for the `click` interface and corresponding explicit function
arguments. (The implicit list of (function) arguments that `click`
passes to the function is already readily available in the context
object.)

As a related problem, the code for checking whether incompatible options
were specified also relies (excessively cleverly) on command-line
argument names and function argument names being equal, and sometimes
conflates short and long option names, leading to it missing some
examples of incompatible options. Therefore, instead, explicitly build
a map of (short and long) option names to option objects and test
whether incompatible option objects were used.