Add finished command-line interface, with tests
Add a finished command-line interface, with interface parity with
vault v0.3.  The implementation has 100% coverage, including the
tests... even if the tests still contain a lot of manual code and data
duplication.  The project settings were updated as well, mostly
expanding the scope of what is covered by type checking and what is
(not) covered by coverage testing.