Fix passphrase-from-SSH-signature calculation
In vault(1), the passphrase derived from an SSH key signature is *not*
just the (framed, binary) SSH signature of the vault UUID.  Instead, the
payload part of the signature, in binary, is converted to base64,
yielding the passphrase.  Rewrite the `Vault.phrase_from_signature`
method to match vault's behavior, and rename it to `phrase_from_key` to
better match the *actual* interface (which takes an SSH key, not
a signature).

(This design – emitting an SSH signature as payload only, in base64 – is
hard-coded into the SSH agent client library used by vault, and likely
not a conscious design by vault's primary author.)