Support textual passphrases, if it is safe
Some level of awareness is necessary to support passphrases stored in
(JSON) config files.  We reject the passphrase if there are multiple
Unicode representations (and thus UTF-8 representations) of the same
text but with different normalizations.  In such a case, a byte string
must be used, and the value cannot currently be stored in the JSON
config file.