Restore __version__ attributes in modules
Version 0.1.0 included broken references to the `__version__` attribute
in other modules, most importantly in the command-line interface.  In
the README, we mention that we support copying the files into the target
directory.  The command-line interface will then have no installed
package to query for its version number.

Fix this by re-duplicating the version number in the `__version__`
attribute, manually, across all top-level modules.

In the long run, we either want Hatchling to support writing the version
back to multiple files, or we want to write a script or hook script that
accomplishes this ourselves, or perhaps we want to publish `sequin` and
`ssh_agent_client` as separate packages so they get their own version
number.  Until then, manually duplicating the number to the top-level
modules is an acceptable compromise to me.