Fix Python package build contents (file selection)
Because `hatch` was using heuristics to determine what to include in the
Python (distribution-)package, some files that were supposed to be
included (the `sequin` and `ssh_agent_client` modules in the `wheel`)
were actually excluded, and some files that were supposed to be excluded
(VCS artifacts from `coverage` and `towncrier`) were actually included.
So, explicitly configure what to include and exclude in the project
settings.