Avoid crashing when overriding key selection on the command-line
A control-flow error resulted in there being a code path where both
`phrase` and `key` arguments were passed to the `derivepassphrase.Vault`
constructor, which does not take a `key` argument.

So, unconditionally discard the `key` argument from the dict of
arguments to pass to the `derivepassphrase.Vault` constructor.  This
also allows streamlining the preceding phrase and key handling logic
somewhat.

For the tests, add further aliases for the second and third dummy keys,
because the comprehensive regression test for this bug needs three
different sources of key selection.