Update documentation to use new module names
Let the documentation reflect the moving and splitting of modules in
26973402aac5d216a66ae3fc8f2d7bcb0fd7ea89.  The filenames are still
mostly the same, to keep links alive where possible.

Because the module list is now more tree-like than list-like, yet still
very small as a whole, abandon the table format in favor of a hierarchy
of (Markdown/HTML) lists.