Use better error message handling in the command-line interface
Instead of blindly relying on `click.Context.fail`, which causes a usage
message to be printed, implement our own error-exit function.

Furthermore, since writing the configuration file is prone to similar
problems as reading it, include a wrapper for writing the configuration
file as well, which ensures errors are properly reported.

Due to the rewrite, some error messages changed slightly in phrasing
and/or in capitalization.