Specify project dependencies more clearly and explicitly
Use proper syntax for dependency version specification (spaces, and
compatible release operator).  If sensible, add a comment as to why we
have this dependency, and why at that version.  Furthermore, for hatch
environments, make use of environment inheritance, and move the
dependencies to the correct environments instead of collecting them all
in the "dev" extra.