Add preliminary tests for the exporter
Includes basic testing of command-line functionality, explicit examples
of each supported format, and proper hatch environment and dependency
selection (with exporter support, or without) to run tests in.  The
major missing pieces are weird-to-reach error conditions in the
internals of the exporter functions (which require invasive mocking) and
explicit API testing of the exporter functions (i.e. beyond each
exporter's "main" function).  Because the internal API is not yet
stable, particularly the v0.2 and v0.3 exporters, this is currently left
as a TODO.

Since we are attempting to support installations without the
cryptography module, we must ensure the exporter modules all stay
importable without triggering import errors.  (This is also relevant for
pytest's test collection phase, where those modules too are scanned for
doctests.)  Currently we stub out each (sub-)module of the cryptography
package implicitly, such that calling any method on the stub raises an
import error, as well as explicitly through a `STUBBED` attribute on the
exporter module.