Test exporter data loading functionality more robustly
Move the `_load_data` function, which deals with determining the correct
vault storage format (with automatic rollover) into the top level, where
the function can be tested in isolation, and stubbed out if necessary.

Also, use the environments with disabled cryptography support to test
for correct failure behavior of the exporter.  Use
`importlib.util.find_spec` instead of an `import` statement to check for
importability of a module, which does not require handling of
`ModuleNotFoundError` if restricted to top-level modules/packages.