Allow all textual strings, but warn on unnormalized ones
Remove the "check if Unicode string has unambiguous byte encoding"
misfeature, which is based on an incorrect understanding of Unicode
normalization forms.  Also remove the `AmbiguousByteRepresentationError`
class.  Now we allow all textual strings as master passphrases.

When storing a (master) passphrase to the configuration or when
interactively generating a passphrase, we emit a warning if we detect
that the master passphrase is not normalized (with the configured
normalization form, stored in `.global.unicode_normalization_form`).