Upgrade ruff to 0.6, and redo configuration and rule selection
hatch currently still defaults to ruff 0.4.5, which includes some
now-deprecated rules and lacks some others.  Furthermore, hatch's
default configuration for ruff is very opinionated and not explained
anywhere.  Upgrading the ruleset to 0.6 still incurs the overhead of
having to look at every single new ruff rule, determining whether it
fits the hatch configuration, and allowing it.  This is a similar amount
of effort to enabling all of ruffs rules and deciding, based on the
flagged violations, whether a rule or rule category is worth keeping.

This has now been implemented: almost all rules are now selected, and
the exceptions are documented in the project settings.  (ruff actually
includes documentation on linting rules that tend to get violated if
ruff is also used for reformatting.)  hatch is configured to use ruff
0.6 and this user configuration, ignoring its own default ruff
configuration.  Some further ruff settings are necessary, e.g. adapting
the default copyright text template to match hatch's copyright text
template, as is one new mypy setting (the "ignore-without-code" error).

Note: the new ruleset has not actually been applied to the codebase yet,
because the effect would be spread out across the codebase, and we
expect multiple topic branches to be merged soon.