Add vault_native exporter function and module docstrings
Add an all-in-one exporter function to the
`derivepassphrase.exporter.vault_native` module, similar to the function
in the `derivepassphrase.exporter.storeroom` module. These two functions
now define the public API of the respective module, and the other
visible names are considered non-public (though documented, because they
explain the inner workings of the respective exporters).  The module
docstrings have been appropriately expanded, and the tests have been
adapted and expanded for the new function.