Rename the configuration file to be subsystem-specific
In preparation for multiple passphrase derivation schemes with separate
settings, use multiple configuration files for each part of the
application ("subsystem").  The only currently established subsystem is
"vault", using `vault.json`, and the existing configuration effectively
corresponds to a "settings" subsystem.

Until v1.0 is released, fall back to `settings.json` if `vault.json`
does not exist. (But include a deprecation warning, and attempt to
migrate to the new filename automatically.)