Reintroduce dynamic versioning via `hatch version`
When we abandoned `hatch version` in
565dd46f8506662fd4abcc8e7f9c7a99ae15958f, we had three top-level
packages sharing a version number, but `hatch version` did not support
reading or updating the version from multiple sources.

Since c4a57f311710768cb18df717a73fd48a8a3077fe however the system is
rooted in the `derivepassphrase` top-level package, and the version
number is internally copied from that package root.  We are therefore
once again in the classic scenario that `hatch version` supports.