Demonstrate config storing and SSH agent use in README
The README now shows a session where the user stores passphrase
derivation settings for a service, and where the user uses an SSH key
for passphrase derivation.

This also makes for a good opportunity to highlight the flaky Windows
support for the SSH agent: because `derivepassphrase` expects to talk to
the SSH agent directly, via a UNIX domain socket, it is unlikely to work
on Windows out of the box because (a) support for UNIX domain sockets is
only available in Windows 10 and higher (ca. 2021), and (b) the two
major SSH agent implementations Pageant (PuTTY) and OpenSSH for Windows
(via the Windows Subsystem for Linux) do not use UNIX domain sockets on
Windows.  See https://github.com/the-13th-letter/issues/13 for details.