Add test fixture for manually spawning known SSH agents
Include pytest fixtures to spawn known SSH agents and interface with
the running SSH agent, in an agent-agnostic way.  Move the key loading
parts from the test functions into the test fixtures.

This generally makes the actual test functions somewhat cleaner and
easier to read, but because the monkeypatch fixture interferes with
these new fixtures, the net improvements to readibility are only
moderate.  The test functions do however profit directly from
reduced copy-and-paste in the key loading part.

Pageant is one of the supported agents, and it behaves markedly
differently than OpenSSH's agent.  In particular, Pageant does not
support adding keys with constraints or re-submitting a key it is
already holding, and *every* key type Pageant currently offers yields
a deterministic signature.  Furthermore, a bug concerning output
buffering in Pageant 0.81 and lower currently makes it impossible to use
Pageant as a subprocess properly without correctly guessing the socket
address.  (This has already been reported upstream.) On the other hand,
Pageant supports ed448 keys, which OpenSSH doesn't.  So, implementing
support for Pageant was very valuable to highlight areas where the code
made unreasonable assumptions about SSH agent behavior, in particular
the availability and behavior of the system SSH agent service.

The new fixtures live in `tests/conftest.py`, following a relevant
pytest convention.  The fixtures themselves are necessarily platform-
and runtime-dependent, so even though they are test code that should be
included in test coverage, all parts dealing with querying the system,
spawning programs, error handling related to the former and ensuring
a certain functionality is available (or skipping the test otherwise)
are excluded from test coverage.  In particular, this includes the
entire fixture to ensure a running agent, and the cleanup part of the
agent with loaded keys fixture; I have however tried various
constellations by hand to ensure the code works if certain agents are
available or unavailable.