Rename reference documentation pages as per the Python standard library convention
A module `spam.ham.eggs` shall be documented in a file
`spam.ham.eggs.html`, or otherwise in `spam.ham.html` or `spam.html` if
its documentation is to be grouped with other sibling modules.  In
particular, documentation is not nested into subdirectories, and the
documentation filename is guessable based on the full module path.

This entails consolidating the `derivepassphrase.exporter` module and
its submodules into the same documentation page.