Switch from towncrier to scriv for changelog management
I have found towncrier to be rather painful when used together with
Markdown entries: it feels very much like a reStructuredText-only tool
retrofitted with patchy Markdown support.  In particular, towncrier
blindly concatenates all entries (assumed to be list items) into the
output (and adds the implied issue links afterwards), manually adjusted
for nicer indentation in an ad-hoc fashion.  This fails horribly if the
changelog entry is anything except a single, paragraph-free list item;
it fails in particular for reference-style links and for multi-paragraph
entries.

So switch to scriv, by Ned Batchelder, which is a similar style to
towncrier.  That too has its issues, such as the automatically
generated, non-configurable HTML4-style heading anchor, but it's
somewhat more tolerant/encouraging of manual editing of the changelog
entries and the changelog master file.

Since we're cleaning up the changelog, use this opportunity to fix the
formatting of the existing entries.