Fix awkward parametrization declaration for SSH agent client tests
Instead of introducing a static method, calling it to obtain test
parameters, then removing it again, use a declaration within the
`pytest.mark.parametrize` decorator arguments, even if that is
supposedly more difficult to read (I wouldn't agree).  Among the
readability benefits (primarily because of compactness), the imperative
version via a static method throws errors on Python 3.9, presumably
because the `staticmethod` object itself is being returned, not the
bound method object, and the former isn't callable in Python 3.9.

On a related note, a future import for annotations was missing as well.