Add a tutorial: setting up three services with master passphrase
The new tutorial is an introductory tutorial that deals with setting up
the first few services with `derivepassphrase vault`.  It deals with
installing the software and configuring the first three services.  I've
included one out-of-scope item (a one-time-password hardware token) and
one ill-specified requirement on special characters as practical
examples of things a new user might actually encounter in practice, and
how to deal with these (or not).

The tutorial is complete content-wise, but probably needs some
copy-editing, and definitely needs some practical feedback.  The
installation section is also untested, and lacking in comprehensiveness,
as there are multiple sane ways to install Python software such as this
one.  Once other documentation is complete, this tutorial may also need
some minor revision on whether and where to link to further topics that
were only briefly touched here.