Align behavior with vault concerning key and phrase in config
When both a key and a passphrase are specified in the vault
configuration, vault(1) would unconditionally use the key, *unless* the
command-line overrides this choice.  `derivepassphrase` however always
gave preference to the most "specific" configuration, and would error
out if both key and passphrase were specified at the same specificity.
While arguably more intuitive, this behavior is a visible deviation from
vault(1), and shall thus be removed.

Besides two instances of the `test_200_is_vault_config` in
`tests.test_derivepassphrase_types`, this also flips the result of
`test_205_service_phrase_if_key_in_global_config` in
`tests.test_derivepassphrase_cli`.  Because that flipped version needs
extra mocking infrastructure – the `sign` function – and because that
mock function already exists in another test (but local to that test),
promote that mock function to global and shift it into the top-level
`tests` module.

Since we had to update the imports in `tests` anyway, we also purged
`dpp.vault...` references in `tests.test_derivepassphrase_cli` in favor
of `vault...`.