Signal and list falsy value cleanup steps that were actually performed
Signal whether cleanup was actually perfomed on the requested object or
not, and if yes, list the actual cleanup steps undertaken.  When
importing a vault configuration on the command-line, issue a warning for
each cleaning step.

Since the warning messages in both the cleanup steps and the check for
non-normalized passphrases report on "paths" in a JSON object, implement
a fully general JSONPath formatting function (single item selection from
the root only).  This harmonizes the warnings output, but also causes
changes in the test cases and expected output.  Additionally, the
JSONPath function name clashes with a common local variable name,
necessitating renaming, and control flow for the validation function and
the vault configuration import action have changed somewhat; the former
to impose a consistent validation order (global first, service-specific
next), the latter to avoid extraneous else-branches.

As a result of all this, this patch is somewhat larger and less concise
than it should be, given the modest magnitude of changes it actually
introduces.