Align behavior with vault concerning config imports
When importing a configuration, the original vault(1) does not replace
its configuration as a whole with the supplied one.  Instead, each
"section" of the new configuration – the global settings, and each named
service's settings – replaces the corresponding section in the old
configuration, if any.  Any sections not mentioned in the new
configuration are left untouched; in particular, existing service
settings are *kept* if they are not mentioned in the new configuration.

`derivepassphrase` uses a simpler, database-like dump/restore model,
overwriting the whole old configuration with the new one.  As
a consequence, old services settings are *dropped* if they are not
mentioned in the new configuration.  This behavior is a visible
deviation from vault(1), and shall thus be removed.

This commit contains the necessary changes to the import machinery to
correctly calculate the new, merged configuration.  Surprisingly, this
already passes all tests, which is more a sign that our tests are
incomplete rather than that the code is robust against this failure
type.  So, in upcoming commits we will introduce functional tests for
the config import/merging machinery second, and bugfixes for other
things we uncovered while writing these functional tests first.