Correctly model vault global and service settings
Excluding extensions due to `derivepassphrase`, in vault(1), the service
settings are a superset of the global settings… though some settings
might be weird in a global setting.  The abstract data types don't
accurately model this, though.  So fix those.

(The sole current `derivepassphrase` extension, the
`unicode_normalization_form` setting, modifies the `phrase` setting, and
so also makes sense to specify globally or locally.  So use the same
inheritance and scoping rules as `phrase`.)

Incidentally, the command-line interface did not respect this
inheritance either… in a more destructive way, asserting that some
options require a service when they don't, or don't take a service when
they do.  Redo the categorization there into three groups instead of
two: service name mandatory, forbidden, or optional (new).