Fix clean up of falsy "length" and "repeat" settings
A falsy length behaves like the standard length, and a falsy repeat
setting forces(!) repetition to be disregarded.  So the former
normalizes to 20 and the latter to 0, regardless of any inherited
setting.  In both cases, we can't just remove the setting, we need to
explicitly set the default value.

(For repeat settings, and upper and lower etc. settings too, there's an
additional twist: the value `0.0` (float) is equivalent to `0` (int),
but the former is a type violation.  This means we need to also make
sure to check the type when normalizing, not just the value.)