Warn the user upon supplying an empty service name
The `derivepassphrase` command-line largely, and vault(1) completely,
treats an empty service name the same as if no service name had been
supplied, switching over to global operation instead of service-specific
operation (or erroring out in case a service must be given).  For
`derivepassphrase`, this is mostly a user interface issue – the
underlying machinery supports empty service names –, but kept for
compatibility with vault(1).  However, this is very easy to diagnose,
and the user would benefit from seeing a warning about a seemingly
omitted service name.  So do that, and issue a warning upon encountering
an empty service name on the command-line or in an imported
configuration.  (The warning message changes slightly in each case.)

Also, add explicit tests for these two scenarios that trigger the
warning.