Add how-to for setting up an SSH key for `derivepassphrase vault`
We explain the necessary software/operating system prerequisites and
configurations.  We further prominently note that Windows is currently
not supported, and that `gpg-agent` behaves differently than other
agents in regard to SSH key management.

We further clean up the list of future work, and ensure styling
consistency with the basic setup (passphrase) tutorial.