Make suitable SSH key listing easier to distinguish
On the one hand, truncate and align the listing as two columns, not
three, by combining key type and (truncated) key data into one column.
For heterogenous lists with different key types, this nicely sets off
the comment column (which the user can change to help distinguish the
keys) from the key data (which the user cannot change).

On the other hand, if truncating the key data for the display, truncate
the *front* of the data, not the back.  For homogenous lists, this
generally leads to better distinguishable key listings: the front
contains information common to all keys (the wire-encoded key type), but
the back contains key-specific information (for RSA, Ed25519 and Ed448
keys at least).