Split the SSH key how-to into how-to and reference documents
Because the original how-to discussed both prerequisites and the how-to
of SSH keys, it was tonally inconsistent.  It makes much more sense from
a reading flow perspective to move the discussion of prerequisites into
a separate reference document and link to it from the how-to page.  So
do exactly that.

Relative to the old how-to page, the new prerequisites reference page
additionally includes sections on how to determine the SSH key type from
the algorithm name used in the wire protocol, and sample transcripts for
generating new SSH keys suitable for `derivepassphrase vault`; this is
based on feedback for the old how-to page.  The new how-to page also
shows the actual key selection dialog instead of only the command-line
to run.