Fix test suite to actually test deterministic signature support
So far, the test suite was silently passing for me, because it requires
either a patched version or a not-yet-released version of PuTTY to
actually run the tests against Pageant (which is the main beneficiary of
deterministic signature detection).  Actually plugging in a suitable
patched Pageant version revealed a couple of key places where we
silently assume that the key type alone determines its suitability for
`derivepassphrase`.  This commit rectifies that.