Decouple deterministic signatures from general SSH agent detection
Instead of tying deterministic signatures directly to the detection of
Pageant specifically, add a general mechanism for attempting to infer
the connected SSH agent from its reported list of extensions.  This
moves the question of *how* we detect certain SSH agents out of the
deterministic signature checking function.

Alas, OpenSSH does not support the extension query message we issue,
despite them supporting the extension system in general *and* stewarding
the SSH agent protocol specification which defines this message
normatively.  So our implementation must tolerate a moderate level of
spec violation.