Let the `running_ssh_agent` test fixture report the agent type
In the current test scenario, where multiple SSH agents are spawned if
possible, it is highly unhelpful to know *that* a running SSH agent
failed, but not *which* agent did.  For debugging purposes, it is better
if the `running_ssh_agent` test fixture reports not only the agent's
socket, but also its type.

It is sufficient to have the type passed as a fixture output/test
function input, because `pytest` will then pretty-print it when a test
function fails.