Reimplement deprecated subcommands, properly, in `click`
Our previous method of dispatching to subcommands was not tying the
supercommand and the subcommand together properly; each command
essentially ran in its own isolated context.  This lead to unnecessary
double-parsing of the command-line arguments, and to `click` becoming
confused as to what the full command-line actually was; in particular,
the help and usage messages gave a wrong command summary.

Instead, we now use custom `click.Group` objects as the supercommands
which default to the "vault" subcommand (with a deprecation warning) if
no explicit subcommand could be matched, rather than throwing an error;
this all happens internally using the normal mechanism for subcommand
dispatch.  Additionally, if the supercommand callback actually is
invoked -- which only happens if the supercommand is given without any
arguments -- we emit a deprecation warning and dispatch manually to the
"vault" subcommand using the same code path as the `click`
implementation would otherwise do.  We assert that the old and new code
paths work via new tests as necessary, particularly for the
"supercommand without arguments" case.

This implementation necessarily messes with some `click` internals,
duplicating the code as it currently stands.  Being a non-public API,
there is of course no guarantee that this will work with future versions
of `click`.  However, we use this solely to implement a compatibility
interface, which should vanish with `derivepassphrase` v1.0 anyway.  The
alternative -- rewriting `derivepassphrase` for another command-line
framework, or rolling our own command-line interface from scratch --
seems even more expensive/wasteful than this solution.