Use the logging system to emit warnings and error messages
Employ the Python standard library module `logging` to issue warning
and error messages, as well as (partially new) debug messages.
Operational warning messages and operational deprecation messages are
handled via the `derivepassphrase` and `derivepassphrase.deprecation`
loggers with a custom handler that prints the resulting formatted
message to standard error.  (Errors, info and debug messages use the
same system, but differ slightly in the formatting.)

Most of the rewriting of warnings and error messages to fit this new
machinery is pretty straight-forward.  The messy parts are the
integration of the `logging` machinery and the `warnings` machinery into
`click`, which isn't really designed with this in mind.  Concretely,
this entails writing a custom logging handler that uses `click.echo`,
a custom formatter that prefixes messages with the program name (and
"warning", "deprecation warning" or "debug" as necessary), a singleton
configuration of this handler and formatter, various wrappers around
standard `click` infrastructure to ensure that this integration code
actually runs during normal operation, and finally, new command-line
options to control the level of output on standard error.